initial calculations for the independent EBGS
These were used in my conversation with Glenn to guide part selection
for the ground station.
I just coppied over the analysis from the tripod weather station,
setting stuff like the leg diameters to zero to adapt the calculations
to work for a stand-alone base plate.

I would like to go back and pretty-up the notebook so that the variable
names make sense for a base plate based GS. I also want to confirm the
dimensions we got from the MFG.

I should figure out why the script is predicting a ballast that's like
1/4 as large as what's suggested by the MFG. Maybe they just have FS=4.